+++ accounting/manual_journal_line.go
@@ -9,10 +9,10 @@ type ManualJournalLine struct {
 	Description string `json:"Description,omitempty" xml:"Description,omitempty"`
 
 	// Net amount of journal line. This will be a positive value for a debit and negative for a credit
-	LineAmount float32 `json:"LineAmount" xml:"LineAmount"`
+	LineAmount float64 `json:"LineAmount" xml:"LineAmount"`
 
 	// The calculated tax amount based on the TaxType and LineAmount
-	TaxAmount float32 `json:"TaxAmount,omitempty" xml:"TaxAmount,omitempty"`
+	TaxAmount float64 `json:"TaxAmount,omitempty" xml:"TaxAmount,omitempty"`
 
 	// Used as an override if the default Tax Code for the selected <AccountCode> is not correct – see TaxTypes.
 	TaxType string `json:"TaxType,omitempty" xml:"TaxType,omitempty"`
