+++ accounting/tax_rate.go
@@ -42,10 +42,10 @@ type TaxRate struct {
 	CanApplyToRevenue bool `json:"CanApplyToRevenue,omitempty" xml:"CanApplyToRevenue,omitempty"`
 
 	// Tax Rate (decimal to 4dp) e.g 12.5000
-	DisplayTaxRate float32 `json:"DisplayTaxRate,omitempty" xml:"DisplayTaxRate,omitempty"`
+	DisplayTaxRate float64 `json:"DisplayTaxRate,omitempty" xml:"DisplayTaxRate,omitempty"`
 
 	// Effective Tax Rate (decimal to 4dp) e.g 12.5000
-	EffectiveRate float32 `json:"EffectiveRate,omitempty" xml:"EffectiveRate,omitempty"`
+	EffectiveRate float64 `json:"EffectiveRate,omitempty" xml:"EffectiveRate,omitempty"`
 }
 
 type TaxRates struct {
